refactor: hhplus_week3 -> hhplus_week3_4_5 패키지명 리팩토링
