+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'hhplus_week3'
+rootProject.name = 'hhplus_week3_4_5'
