--- src/main/java/com/hhplus/hhplus_week3/HhplusWeek3Application.java
@@ -1,13 +1,0 @@
-package com.hhplus.hhplus_week3;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HhplusWeek3Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HhplusWeek3Application.class, args);
-    }
-
-}
