--- src/main/java/com/hhplus/hhplus_week3/application/controllers/cart/dto/AddCartApiReqDto.java
@@ -1,10 +1,0 @@
-package com.hhplus.hhplus_week3.application.controllers.cart.dto;
-
-import java.io.Serializable;
-
-public record AddCartApiReqDto(
-    Long productId,
-    Long productOptionId,
-    int buyCnt
-) implements Serializable {
-}
