+++ src/main/java/com/hhplus/hhplus_week3_4_5/HhplusWeek345Application.java
@@ -1,0 +1,13 @@
+package com.hhplus.hhplus_week3_4_5;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HhplusWeek345Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HhplusWeek345Application.class, args);
+    }
+
+}
