+++ src/main/java/com/hhplus/hhplus_week3_4_5/application/domain/product/ProductEnums.java
@@ -1,0 +1,24 @@
+package com.hhplus.hhplus_week3_4_5.application.domain.product;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+public class ProductEnums {
+    @Getter
+    @RequiredArgsConstructor
+    public enum Type {
+        SINGLE, OPTION
+    }
+
+    @Getter
+    @RequiredArgsConstructor
+    public enum OptionType {
+        BASIC, ADD
+    }
+
+    @Getter
+    @RequiredArgsConstructor
+    public enum Ranking {
+        POPULAR, RECENT, HIGH, LOW
+    }
+}
