--- src/test/java/com/hhplus/hhplus_week3/HhplusWeek3ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.hhplus.hhplus_week3;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HhplusWeek3ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
