+++ src/test/java/com/hhplus/hhplus_week3_4_5/HhplusWeek345ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.hhplus.hhplus_week3_4_5;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HhplusWeek345ApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
